backend/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.12.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
class-registry==2.1.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.16
django-cors-headers==3.11.0
django-filter==2.2.0
django-import-export==2.5.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-filters==0.10.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
et-xmlfile==1.1.0
filters==1.3.2
flake8==5.0.4
flake8-docstrings==1.6.0
flake8-isort==5.0.3
gunicorn==20.0.4
idna==3.4
importlib-metadata==4.2.0
iniconfig==1.1.1
ipython==7.34.0
isort==5.11.4
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==8.2.0
mypy==0.971
mypy-extensions==0.4.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.0
package-name==0.1
packaging==22.0
parso==0.8.3
pathspec==0.10.3
pep517==0.13.0
pickleshare==0.7.5
Pillow==8.3.1
pip-tools==6.10.0
pitest==0.1
platformdirs==2.6.0
pluggy==0.13.1
prompt-toolkit==3.0.31
psycopg2==2.9.5
psycopg2-binary==2.9.5
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytils==0.4.1
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
reportlab==3.6.8
regex==2022.10.31
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
somepackage==1.2.3
sqlparse==0.3.1
tablib==3.3.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.4.0
typed-ast==1.5.4
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.1.9
xlrd==2.0.1
xlwt==1.3.0
zipp==3.11.0